mqtt: Fix title for webradio
Starting with commit 751f182 the webradio stream title gets appended to
the station title. This introduced a bug and now all further stream
titles are simply appended.

Fix the problem by simply replacing the title with the stream title
updates. That way the station name is no longer visible after a while,
but that behavior is aligned with titles on the SD card, where we
also don't differentiate between the track title / artist / album / ...